fix(recorder): Remove recoder web page
