--- frappe/www/recorder.html
@@ -1,11 +1,0 @@
-{% extends "templates/web.html" %}
-
-{% block title %}{{ heading or "Recorder"}}{% endblock %}
-
-{% block page_content %}
-
-<div id="recorder"></div>
-<script type="text/javascript" src="/assets/js/frappe-recorder.min.js"></script>
-<style> {{ highlight }} </style>
-
-{% endblock %}
